fix: move chatroom to chat directory
